+++ scripts/build.ts
@@ -26,58 +26,42 @@ async function getArgs() {
 }
 
 async function extractIconsIfNecessary() {
-  if (fs.existsSync('./dist/icons')) {
+  if (fs.existsSync('./dist/icons') && fs.existsSync('./src/icons.ts')) {
     return;
   }
 
-  console.info('Extracting SVG Icons');
-  console.time('Bundle Lazy Icons');
+  console.info('Extracting SVG Lazy Icons');
+  console.time('Lazy Icons');
   fs.mkdirSync('./dist/icons', { recursive: true });
 
-  const promises = [
-    import('react-icons/ai'),
-    import('react-icons/bi'),
-    import('react-icons/bs'),
-    import('react-icons/cg'),
-    import('react-icons/ci'),
-    import('react-icons/di'),
-    import('react-icons/fa'),
-    import('react-icons/fa6'),
-    import('react-icons/fc'),
-    import('react-icons/fi'),
-    import('react-icons/gi'),
-    import('react-icons/go'),
-    import('react-icons/gr'),
-    import('react-icons/hi'),
-    import('react-icons/hi2'),
-    import('react-icons/im'),
-    import('react-icons/io'),
-    import('react-icons/io5'),
-    import('react-icons/lia'),
-    import('react-icons/lu'),
-    import('react-icons/md'),
-    import('react-icons/pi'),
-    import('react-icons/ri'),
-    import('react-icons/rx'),
-    import('react-icons/si'),
-    import('react-icons/sl'),
-    import('react-icons/tb'),
-    import('react-icons/tfi'),
-    import('react-icons/ti'),
-    import('react-icons/vsc'),
-    import('react-icons/wi'),
-  ];
+  let tsFile = '// This file is generated on build, do not edit.\nexport type IconName =';
+  const entries = fs.readdirSync('./node_modules/react-icons');
 
-  let families = await Promise.all(promises);
-  for (const family of families) {
+  for (const entry of entries) {
+    const entryPath = path.join('./node_modules/react-icons', entry);
+    const isDir = fs.statSync(entryPath).isDirectory();
+
+    if (!isDir || entry === 'lib') {
+      continue;
+    }
+
+    console.info('Extracting icon family:', entry);
+
+    const family = await import(`react-icons/${entry}`);
     for (const [name, ElementConstructor] of Object.entries(family)) {
+      if (typeof ElementConstructor !== 'function') {
+        continue;
+      }
       const element = ElementConstructor({ size: '1em' });
       const svg = renderToStaticMarkup(element);
       fs.writeFileSync(`./dist/icons/${name}.svg`, svg);
+      tsFile += `\n  | '${name}'`;
     }
   }
 
-  console.timeEnd('Bundle Lazy Icons');
+  tsFile += ';\n';
+  fs.writeFileSync('./src/icons.ts', tsFile);
+  console.timeEnd('Lazy Icons');
 }
 
 const appFolders = fs
@@ -168,7 +152,13 @@ function startDevServer() {
     loader: {
       '.yml': 'text',
     },
-    plugins: [CssModulesPlugin({ localsConvention: 'camelCase', pattern: 'do-not-use-on-themes-[local]-[hash]' }), OwnPlugin],
+    plugins: [
+      CssModulesPlugin({
+        localsConvention: 'camelCase',
+        pattern: 'do-not-use-on-themes-[local]-[hash]',
+      }),
+      OwnPlugin,
+    ],
   });
 
   if (serve) {