+++ src/apps/shared/components/Icon/index.tsx
@@ -1,3 +1,4 @@
+import { IconName } from '@icons';
 import { HTMLAttributes } from 'react';
 
 import { cx } from '../../styles';
@@ -5,7 +6,7 @@ import InlineSVG from '../InlineSvg';
 import cs from './index.module.css';
 
 interface ReactIconProps extends HTMLAttributes<HTMLElement> {
-  iconName: string;
+  iconName: IconName;
   size?: string | number;
   color?: string;
 }