+++ src/background/modules/bluetooth/infrastructure.rs
@@ -20,6 +20,8 @@ use crate::{
 use crate::error_handler::Result;
 
 pub fn register_bluetooth_events() {
+    log_error!(trace_lock!(BLUETOOTH_MANAGER).register_for_bt_devices());
+
     BluetoothManager::subscribe(|event| match event {
         BluetoothEvent::DevicesChanged(items) => {
             log_error!(get_app_handle().emit(
@@ -38,9 +40,6 @@ pub fn register_bluetooth_events() {
                     .map_into()
                     .collect::<Vec<BluetoothDevice>>()
             ));
-        }
-        BluetoothEvent::RadioStateChanged(state) => {
-            log_error!(get_app_handle().emit(SeelenEvent::BluetoothRadioStateChanged, state));
         }
     });
     BluetoothPairManager::subscribe(|event| match event {
@@ -78,14 +77,12 @@ pub fn get_connected_bluetooth_devices() -> Result<Vec<BluetoothDevice>> {
 
 #[tauri::command(async)]
 pub fn get_bluetooth_radio_state() -> Result<bool> {
-    let manager = trace_lock!(BLUETOOTH_MANAGER);
-    manager.get_radio_state()
+    Ok(true)
 }
 
 #[tauri::command(async)]
-pub fn set_bluetooth_radio_state(state: bool) -> Result<()> {
-    let manager = trace_lock!(BLUETOOTH_MANAGER);
-    manager.set_radio_state(state)
+pub fn set_bluetooth_radio_state(_state: bool) -> Result<()> {
+    Ok(())
 }
 
 #[tauri::command(async)]